+++ src/lib.rs
@@ -6,5 +6,5 @@ pub mod pathlib;
 pub mod string;
 pub mod time;
 
-pub use crate::enumerate::Enumerator;
+pub use crate::enumerate::Iterable;
 pub use crate::input::input;